requirements.txt
django>=5.2.3
djangorestframework>=3.16.0
pandas>=2.3.0
openpyxl>=3.1.5 